function-source/requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.*
requests==2.*
cloudevents==1.*
